soft-delete feature 'setDeleted' configuration option
soft delete can now take a string or an object as configuration,
exposing an additional field: setDeleted

setDeleted allows you to specify a function that returns a value that
will be used when a record is soft deleted

use case examples would be a date object or the id of a user that
triggered the deletion.

setDeleted checks to ensure the value returned is truthy, and may be
asynchronous.

Other notes:

-tests refactored, no longer throw promise warnings.

- soft-delete now tries to look for a custom authentication entity

- field before defaulting to 'user' in the 'throwIfDeleted' subroutine.

- removed deprecated babel-preset-es2015 in favour of babel-preset-env.
I'm aware that feathers.js is planning on removing babel, but for the
time being this prevents another deprecation warning.

- removed babel-polyfill in favour of babel-transform-runtime.
babel-polyfill pollutes the global namespace whereas
babel-transform-runtime is intended for libraries.

- I also fixed some populate tests that were failling that had nothing
to do with my changes.