Prevents updating a value with the preventChange hook using a falsy value
Currently it is possible to modify a value despite the presence of the preventChange hook because no check is made against negative values, such as 0, false or null for example. This can potentially be a big security hole, as it is possible to remove any value that would normally be protected. By using the object.hasOwnProperty method, we ensure that none of these values can be filled in.